+++ gluten-core/src/main/java/org/apache/gluten/memory/memtarget/MemoryTargetUtil.java
