+++ gluten-core/src/main/java/org/apache/gluten/task/TaskListener.scala
@@ -14,18 +14,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.spark.util
+package org.apache.gluten.task
 
-/**
- * Manages the lifecycle for a specific type of memory resource managed by Spark. See also
- * `org.apache.spark.util.TaskResources`.
- */
-trait TaskResource {
-  @throws(classOf[Exception])
-  def release(): Unit
+import org.apache.spark.TaskFailedReason
 
-  // #release() will be called in higher precedence if the manager has higher priority
-  def priority(): Int = 100
-
-  def resourceName(): String
+trait TaskListener {
+  def onTaskStart(): Unit
+  def onTaskSucceeded(): Unit
+  def onTaskFailed(failureReason: TaskFailedReason): Unit
 }
